requirements.txt
appdirs==1.4.4
argcomplete==1.12.3
Pillow==8.3.1
py-graphql-client==0.1.1
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==5.4.1
qrcode==7.3
six==1.16.0
websocket-client==0.54.0
